Move IGeneratorParser into feature package
