+++ src/main/java/cofh/cofhworld/feature/parser/DecorationParser.java
@@ -1,6 +1,6 @@
 package cofh.cofhworld.feature.parser;
 
-import cofh.cofhworld.decoration.IGeneratorParser;
+import cofh.cofhworld.feature.IGeneratorParser;
 import cofh.cofhworld.init.FeatureParser;
 import cofh.cofhworld.util.WeightedRandomBlock;
 import cofh.cofhworld.world.generator.WorldGenDecoration;
